:sparkles: implement edit for text-field page and saving details more regularly
