+++ js/details/modal/json.js
@@ -1,3 +1,6 @@
+import { saveCurrentDetailSetup } from "../details-functions.js";
+import { downloadArea, uploadArea } from "../../components/upload-download.js";
+
 function setUpJSONDownloadUpload(id) {
     // Custom Filename
     downloadArea(id, "custom-details", () => downloadJSON(id), ".json");
@@ -21,88 +24,12 @@ function downloadJSON(id) {
                 .select(".download-name")
                 .attr("placeholder") + ".json";
     }
-    var details = getDetails("details");
-
-    // based on select2, reorder and tag with hidden
-    var newDetails = [];
-    d3.select("#reorder-columns")
-        .selectAll("td")
-        .each(function() {
-            let detail = _.find(details, {
-                title: d3
-                    .select(this)
-                    .select("span")
-                    .text(),
-            });
-            if (
-                d3
-                    .select(this)
-                    .select("i")
-                    .attr("class") === "bi bi-eye-slash-fill"
-            ) {
-                detail["hidden"] = true;
-            }
-            // custom saves for each
-            if (!detail.hidden && detail.id) {
-                var d = d3.select("#details").select("#" + detail.id);
-                if (!d.empty()) {
-                    switch (detail.type) {
-                        case "team":
-                            // save teams
-                            detail.blueTeamName = d
-                                .select("#blue-team-name")
-                                .property("value");
-                            detail.orangeTeamName = d
-                                .select("#orange-team-name")
-                                .property("value");
-                            detail.checked = d3
-                                .select("input[name='team-bool']:checked")
-                                .property("id");
-                            break;
-
-                        case "player":
-                        case "text-field":
-                            // save current entry
-                            detail["defaultValue"] = d
-                                .select("input")
-                                .property("value");
-                            break;
-
-                        case "shot-type":
-                        case "dropdown":
-                            // save currently selected option
-                            let selectedValue = d
-                                .select("select")
-                                .property("value");
-                            detail.options = detail.options.map(function(o) {
-                                let option = { value: o.value };
-                                if (o.value === selectedValue) {
-                                    option.selected = true;
-                                }
-                                return option;
-                            });
-                            break;
-
-                        case "radio":
-                            // save current selection
-                            let checkedValue = d
-                                .select(`input[name='${detail.id}']:checked`)
-                                .property("value");
-                            detail.options = detail.options.map(function(o) {
-                                let option = { value: o.value };
-                                if (o.value === checkedValue) {
-                                    option.checked = true;
-                                }
-                                return option;
-                            });
-                            break;
-                    }
-                }
-            }
-            newDetails.push(detail);
-        });
-
-    download(JSON.stringify(newDetails), fileName, "application/json");
+    saveCurrentDetailSetup();
+    download(
+        JSON.stringify(getDetails(), null, 2),
+        fileName,
+        "application/json"
+    );
 }
 
 function uploadJSON(id, uploadId, e) {
@@ -129,3 +56,5 @@ function uploadJSON(id, uploadId, e) {
             .attr("class", "form-control is-invalid");
     }
 }
+
+export { setUpJSONDownloadUpload };
